+++ lib/stash/wrapper/inventory.rb
@@ -14,8 +14,18 @@ module Stash
       # creates a new {Inventory} object
       # @param files [List<StashFile>] The inventory of files
       def initialize(files:)
-        self.files = files
+        self.files = valid_file_array(files)
         self.num_files = files.size
+      end
+
+      private
+
+      def valid_file_array(files)
+        fail ArgumentError, "specified file list does not appear to be an array of StashFiles: #{files.inspect}" unless files.is_a?(Array)
+        files.each_with_index do |f, i|
+          fail ArgumentError, "files[#{i}] does not appear to be a StashFile: #{f.inspect}" unless f.is_a?(StashFile)
+        end
+        files
       end
     end
   end