+++ CHANGES.md
@@ -4,6 +4,7 @@
   - `StashWrapper`
   - `StashAdministrative` (fixes issue #2)
   - `Embargo`
+  - `Identifier`
   - `Version`
 
 # 0.1.9 (18 May 2016)