+++ CHANGES.md
@@ -9,6 +9,7 @@
   - `Embargo`
   - `Identifier`
   - `Inventory`
+  - `Size`
   - `Version`
 
 # 0.1.9 (18 May 2016)