+++ CHANGES.md
@@ -1,3 +1,7 @@
+## 0.1.1
+
+- `logger` is now a parameter passed to `Stash::Sword::Client` instead of a global singleton.
+
 ## 0.1.0 (8 June 2016)
 
 - Initial release.